+++ src/WorkerHandler.js
@@ -1,25 +1,21 @@
-'use strict';
+"use strict";
 
-var Promise = require('./Promise');
-var environment = require('./environment');
-var requireFoolWebpack = require('./requireFoolWebpack');
+var Promise = require("./Promise");
+var environment = require("./environment");
+var requireFoolWebpack = require("./requireFoolWebpack");
 
 /**
  * Special message sent by parent which causes a child process worker to terminate itself.
  * Not a "message object"; this string is the entire message.
  */
-var TERMINATE_METHOD_ID = '__workerpool-terminate__';
-
-/**
- * If sending `TERMINATE_METHOD_ID` does not cause the child process to exit in this many milliseconds,
- * force-kill the child process.
- */
-var CHILD_PROCESS_EXIT_TIMEOUT = 1000;
+var TERMINATE_METHOD_ID = "__workerpool-terminate__";
 
 function ensureWorkerThreads() {
-  var WorkerThreads = tryRequireWorkerThreads()
+  var WorkerThreads = tryRequireWorkerThreads();
   if (!WorkerThreads) {
-    throw new Error('WorkerPool: workerType = \'thread\' is not supported, Node >= 11.7.0 required')
+    throw new Error(
+      "WorkerPool: workerType = 'thread' is not supported, Node >= 11.7.0 required"
+    );
   }
 
   return WorkerThreads;
@@ -28,16 +24,24 @@ function ensureWorkerThreads() {
 // check whether Worker is supported by the browser
 function ensureWebWorker() {
   // Workaround for a bug in PhantomJS (Or QtWebkit): https://github.com/ariya/phantomjs/issues/14534
-  if (typeof Worker !== 'function' && (typeof Worker !== 'object' || typeof Worker.prototype.constructor !== 'function')) {
-    throw new Error('WorkerPool: Web Workers not supported');
+  if (
+    typeof Worker !== "function" &&
+    (typeof Worker !== "object" ||
+      typeof Worker.prototype.constructor !== "function")
+  ) {
+    throw new Error("WorkerPool: Web Workers not supported");
   }
 }
 
 function tryRequireWorkerThreads() {
   try {
-    return requireFoolWebpack('worker_threads');
+    return requireFoolWebpack("worker_threads");
   } catch (error) {
-    if (typeof error === 'object' && error !== null && error.code === 'MODULE_NOT_FOUND') {
+    if (
+      typeof error === "object" &&
+      error !== null &&
+      error.code === "MODULE_NOT_FOUND"
+    ) {
       // no worker_threads available (old version of node.js)
       return null;
     } else {
@@ -48,45 +52,62 @@ function tryRequireWorkerThreads() {
 
 // get the default worker script
 function getDefaultWorker() {
-  if (environment.platform === 'browser') {
+  if (environment.platform === "browser") {
     // test whether the browser supports all features that we need
-    if (typeof Blob === 'undefined') {
-      throw new Error('Blob not supported by the browser');
+    if (typeof Blob === "undefined") {
+      throw new Error("Blob not supported by the browser");
     }
-    if (!window.URL || typeof window.URL.createObjectURL !== 'function') {
-      throw new Error('URL.createObjectURL not supported by the browser');
+    if (!window.URL || typeof window.URL.createObjectURL !== "function") {
+      throw new Error("URL.createObjectURL not supported by the browser");
     }
 
     // use embedded worker.js
-    var blob = new Blob([require('./generated/embeddedWorker')], { type: 'text/javascript' });
+    var blob = new Blob([require("./generated/embeddedWorker")], {
+      type: "text/javascript",
+    });
     return window.URL.createObjectURL(blob);
-  }
-  else {
+  } else {
     // use external worker.js in current directory
-    return __dirname + '/worker.js';
+    return __dirname + "/worker.js";
   }
 }
 
 function setupWorker(script, options) {
-  if (options.workerType === 'web') { // browser only
+  if (options.workerType === "web") {
+    // browser only
     ensureWebWorker();
     return setupBrowserWorker(script, Worker);
-  } else if (options.workerType === 'thread') { // node.js only
+  } else if (options.workerType === "thread") {
+    // node.js only
     WorkerThreads = ensureWorkerThreads();
-    return setupWorkerThreadWorker(script, WorkerThreads);
-  } else if (options.workerType === 'process' || !options.workerType) { // node.js only
-    return setupProcessWorker(script, resolveForkOptions(options), requireFoolWebpack('child_process'));
-  } else { // options.workerType === 'auto' or undefined
-    if (environment.platform === 'browser') {
+    return setupWorkerThreadWorker(
+      script,
+      WorkerThreads,
+      options.workerThreadOpts
+    );
+  } else if (options.workerType === "process" || !options.workerType) {
+    // node.js only
+    return setupProcessWorker(
+      script,
+      resolveForkOptions(options),
+      requireFoolWebpack("child_process")
+    );
+  } else {
+    // options.workerType === 'auto' or undefined
+    if (environment.platform === "browser") {
       ensureWebWorker();
       return setupBrowserWorker(script, Worker);
-    }
-    else { // environment.platform === 'node'
+    } else {
+      // environment.platform === 'node'
       var WorkerThreads = tryRequireWorkerThreads();
       if (WorkerThreads) {
         return setupWorkerThreadWorker(script, WorkerThreads);
       } else {
-        return setupProcessWorker(script, resolveForkOptions(options), requireFoolWebpack('child_process'));
+        return setupProcessWorker(
+          script,
+          resolveForkOptions(options),
+          requireFoolWebpack("child_process")
+        );
       }
     }
   }
@@ -103,21 +124,21 @@ function setupBrowserWorker(script, Worker) {
       callback(message.data);
     });
   };
-  worker.send = function (message) {
-    this.postMessage(message);
+  worker.send = function (message, transfer) {
+    this.postMessage(message, transfer);
   };
   return worker;
 }
 
-function setupWorkerThreadWorker(script, WorkerThreads) {
+function setupWorkerThreadWorker(script, WorkerThreads, workerThreadOptions) {
   var worker = new WorkerThreads.Worker(script, {
     stdout: false, // automatically pipe worker.STDOUT to process.STDOUT
-    stderr: false  // automatically pipe worker.STDERR to process.STDERR
+    stderr: false, // automatically pipe worker.STDERR to process.STDERR
+    ...workerThreadOptions,
   });
   worker.isWorkerThread = true;
-  // make the worker mimic a child_process
-  worker.send = function (message) {
-    this.postMessage(message);
+  worker.send = function (message, transfer) {
+    this.postMessage(message, transfer);
   };
 
   worker.kill = function () {
@@ -134,11 +155,13 @@ function setupWorkerThreadWorker(script, WorkerThreads) {
 
 function setupProcessWorker(script, options, child_process) {
   // no WorkerThreads, fallback to sub-process based workers
-  var worker = child_process.fork(
-    script,
-    options.forkArgs,
-    options.forkOpts
-  );
+  var worker = child_process.fork(script, options.forkArgs, options.forkOpts);
+
+  // ignore transfer argument since it is not supported by process
+  var send = worker.send;
+  worker.send = function (message) {
+    return send.call(worker, message);
+  };
 
   worker.isChildProcess = true;
   return worker;
@@ -148,31 +171,32 @@ function setupProcessWorker(script, options, child_process) {
 function resolveForkOptions(opts) {
   opts = opts || {};
 
-  var processExecArgv = process.execArgv.join(' ');
-  var inspectorActive = processExecArgv.indexOf('--inspect') !== -1;
-  var debugBrk = processExecArgv.indexOf('--debug-brk') !== -1;
+  var processExecArgv = process.execArgv.join(" ");
+  var inspectorActive = processExecArgv.indexOf("--inspect") !== -1;
+  var debugBrk = processExecArgv.indexOf("--debug-brk") !== -1;
 
   var execArgv = [];
   if (inspectorActive) {
-    execArgv.push('--inspect=' + opts.debugPort);
+    execArgv.push("--inspect=" + opts.debugPort);
 
     if (debugBrk) {
-      execArgv.push('--debug-brk');
+      execArgv.push("--debug-brk");
     }
   }
 
   process.execArgv.forEach(function (arg) {
-    if (arg.indexOf('--max-old-space-size') > -1) {
-      execArgv.push(arg)
+    if (arg.indexOf("--max-old-space-size") > -1) {
+      execArgv.push(arg);
     }
-  })
+  });
 
   return Object.assign({}, opts, {
     forkArgs: opts.forkArgs,
     forkOpts: Object.assign({}, opts.forkOpts, {
-      execArgv: (opts.forkOpts && opts.forkOpts.execArgv || [])
-        .concat(execArgv)
-    })
+      execArgv: ((opts.forkOpts && opts.forkOpts.execArgv) || []).concat(
+        execArgv
+      ),
+    }),
   });
 }
 
@@ -182,14 +206,14 @@ function resolveForkOptions(opts) {
  * @return {Error} The equivalent Error.
  */
 function objectToError(obj) {
-  var temp = new Error('')
-  var props = Object.keys(obj)
+  var temp = new Error("");
+  var props = Object.keys(obj);
 
   for (var i = 0; i < props.length; i++) {
-    temp[props[i]] = obj[props[i]]
+    temp[props[i]] = obj[props[i]];
   }
 
-  return temp
+  return temp;
 }
 
 /**
@@ -208,6 +232,10 @@ function WorkerHandler(script, _options) {
   this.worker = setupWorker(this.script, options);
   this.debugPort = options.debugPort;
   this.maxJobsPerWorker = options.maxJobsPerWorker || Infinity;
+  this.forkOpts = options.forkOpts;
+  this.forkArgs = options.forkArgs;
+  this.workerThreadOpts = options.workerThreadOpts;
+  this.workerTerminateTimeout = options.workerTerminateTimeout;
 
   // The ready message is only sent if the worker.add method is called (And the default script is not used)
   if (!script) {
@@ -216,11 +244,11 @@ function WorkerHandler(script, _options) {
 
   // queue for requests that are received before the worker is ready
   this.requestQueue = [];
-  this.worker.on('message', function (response) {
+  this.worker.on("message", function (response) {
     if (me.terminated) {
       return;
     }
-    if (typeof response === 'string' && response === 'ready') {
+    if (typeof response === "string" && response === "ready") {
       me.worker.ready = true;
       dispatchQueuedRequests();
     } else {
@@ -229,7 +257,7 @@ function WorkerHandler(script, _options) {
       var task = me.processing[id];
       if (task !== undefined) {
         if (response.isEvent) {
-          if (task.options && typeof task.options.on === 'function') {
+          if (task.options && typeof task.options.on === "function") {
             task.options.on(response.payload);
           }
         } else {
@@ -245,8 +273,7 @@ function WorkerHandler(script, _options) {
           // resolve the task's promise
           if (response.error) {
             task.resolver.reject(objectToError(response.error));
-          }
-          else {
+          } else {
             task.resolver.resolve(response.result);
           }
         }
@@ -269,25 +296,25 @@ function WorkerHandler(script, _options) {
   // send all queued requests to worker
   function dispatchQueuedRequests() {
     for (const request of me.requestQueue.splice(0)) {
-      me.worker.send(request);
+      me.worker.send(request.message, request.transfer);
     }
   }
 
   var worker = this.worker;
   // listen for worker messages error and exit
-  this.worker.on('error', onError);
-  this.worker.on('exit', function (exitCode, signalCode) {
-    var message = 'Workerpool Worker terminated Unexpectedly\n';
+  this.worker.on("error", onError);
+  this.worker.on("exit", function (exitCode, signalCode) {
+    var message = "Workerpool Worker terminated Unexpectedly\n";
 
-    message += '    exitCode: `' + exitCode + '`\n';
-    message += '    signalCode: `' + signalCode + '`\n';
+    message += "    exitCode: `" + exitCode + "`\n";
+    message += "    signalCode: `" + signalCode + "`\n";
 
-    message += '    workerpool.script: `' + me.script + '`\n';
-    message += '    spawnArgs: `' + worker.spawnargs + '`\n';
-    message += '    spawnfile: `' + worker.spawnfile + '`\n'
+    message += "    workerpool.script: `" + me.script + "`\n";
+    message += "    spawnArgs: `" + worker.spawnargs + "`\n";
+    message += "    spawnfile: `" + worker.spawnfile + "`\n";
 
-    message += '    stdout: `' + worker.stdout + '`\n'
-    message += '    stderr: `' + worker.stderr + '`\n'
+    message += "    stdout: `" + worker.stdout + "`\n";
+    message += "    stderr: `" + worker.stderr + "`\n";
 
     onError(new Error(message));
   });
@@ -296,6 +323,7 @@ function WorkerHandler(script, _options) {
 
   this.terminating = false;
   this.terminated = false;
+  this.cleaning = false;
   this.terminationHandler = null;
   this.lastId = 0;
   this.handledJobCount = 0;
@@ -306,7 +334,7 @@ function WorkerHandler(script, _options) {
  * @return {Promise.<String[], Error>} methods
  */
 WorkerHandler.prototype.methods = function () {
-  return this.exec('methods');
+  return this.exec("methods");
 };
 
 /**
@@ -329,28 +357,31 @@ WorkerHandler.prototype.exec = function (method, params, resolver, options) {
   this.processing[id] = {
     id: id,
     resolver: resolver,
-    options: options
+    options: options,
   };
 
   // build a JSON-RPC request
   var request = {
-    id: id,
-    method: method,
-    params: params
+    message: {
+      id: id,
+      method: method,
+      params: params,
+    },
+    transfer: options && options.transfer,
   };
 
   if (this.exhausted()) {
-    resolver.reject(new Error('Worker is exhausted'));
+    resolver.reject(new Error("Worker is exhausted"));
   } else {
     // Increment the handled job counter.
     ++this.handledJobCount;
   }
 
   if (this.terminated) {
-    resolver.reject(new Error('Worker is terminated'));
+    resolver.reject(new Error("Worker is terminated"));
   } else if (this.worker.ready) {
     // send the request to the worker
-    this.worker.send(request);
+    this.worker.send(request.message, request.transfer);
   } else {
     this.requestQueue.push(request);
   }
@@ -358,30 +389,35 @@ WorkerHandler.prototype.exec = function (method, params, resolver, options) {
   // on cancellation, force the worker to terminate
   var me = this;
   return resolver.promise.catch(function (error) {
-    if (error instanceof Promise.CancellationError || error instanceof Promise.TimeoutError) {
+    if (
+      error instanceof Promise.CancellationError ||
+      error instanceof Promise.TimeoutError
+    ) {
       // remove this task from the queue. It is already rejected (hence this
       // catch event), and else it will be rejected again when terminating
       delete me.processing[id];
 
       // terminate worker
-      return me.terminateAndNotify(true)
-        .then(function () {
+      return me.terminateAndNotify(true).then(
+        function () {
           throw error;
-        }, function (err) {
+        },
+        function (err) {
           throw err;
-        });
+        }
+      );
     } else {
       throw error;
     }
-  })
+  });
 };
 
 /**
- * Test whether the worker is working or not
+ * Test whether the worker is processing any tasks or cleaning up before termination.
  * @return {boolean} Returns true if the worker is busy
  */
 WorkerHandler.prototype.busy = function () {
-  return Object.keys(this.processing).length > 0;
+  return this.cleaning || Object.keys(this.processing).length > 0;
 };
 
 /**
@@ -406,22 +442,23 @@ WorkerHandler.prototype.terminate = function (force, callback) {
     // cancel all tasks in progress
     for (var id in this.processing) {
       if (this.processing[id] !== undefined) {
-        this.processing[id].resolver.reject(new Error('Worker terminated'));
+        this.processing[id].resolver.reject(new Error("Worker terminated"));
       }
     }
     this.processing = Object.create(null);
   }
 
-  if (typeof callback === 'function') {
+  if (typeof callback === "function") {
     this.terminationHandler = callback;
   }
   if (!this.busy()) {
     // all tasks are finished. kill the worker
     var cleanup = function (err) {
       me.terminated = true;
+      me.cleaning = false;
       if (me.worker != null && me.worker.removeAllListeners) {
         // removeAllListeners is only available for child_process
-        me.worker.removeAllListeners('message');
+        me.worker.removeAllListeners("message");
       }
       me.worker = null;
       me.terminating = false;
@@ -430,54 +467,49 @@ WorkerHandler.prototype.terminate = function (force, callback) {
       } else if (err) {
         throw err;
       }
-    }
+    };
 
     if (this.worker) {
-      if (typeof this.worker.kill === 'function') {
+      if (typeof this.worker.kill === "function") {
         if (this.worker.killed) {
-          cleanup(new Error('worker already killed!'));
+          cleanup(new Error("worker already killed!"));
           return;
         }
 
-        if (this.worker.isChildProcess) {
-          var cleanExitTimeout = setTimeout(function () {
-            if (me.worker) {
-              me.worker.kill();
-            }
-          }, CHILD_PROCESS_EXIT_TIMEOUT);
-
-          this.worker.once('exit', function () {
-            clearTimeout(cleanExitTimeout);
-            if (me.worker) {
-              me.worker.killed = true;
-            }
-            cleanup();
-          });
-
-          if (this.worker.ready) {
-            this.worker.send(TERMINATE_METHOD_ID);
-          } else {
-            this.worker.requestQueue.push(TERMINATE_METHOD_ID)
+        // child process and worker threads
+        var cleanExitTimeout = setTimeout(function () {
+          if (me.worker) {
+            me.worker.kill();
           }
-        } else {
-          // worker_thread
-          this.worker.kill();
-          this.worker.killed = true;
+        }, this.workerTerminateTimeout);
+
+        this.worker.once("exit", function () {
+          clearTimeout(cleanExitTimeout);
+          if (me.worker) {
+            me.worker.killed = true;
+          }
           cleanup();
+        });
+
+        if (this.worker.ready) {
+          this.worker.send(TERMINATE_METHOD_ID);
+        } else {
+          this.requestQueue.push({ message: TERMINATE_METHOD_ID });
         }
+
+        // mark that the worker is cleaning up resources
+        // to prevent new tasks from being executed
+        this.cleaning = true;
         return;
-      }
-      else if (typeof this.worker.terminate === 'function') {
+      } else if (typeof this.worker.terminate === "function") {
         this.worker.terminate(); // web worker
         this.worker.killed = true;
-      }
-      else {
-        throw new Error('Failed to terminate worker');
+      } else {
+        throw new Error("Failed to terminate worker");
       }
     }
     cleanup();
-  }
-  else {
+  } else {
     // we can't terminate immediately, there are still tasks being executed
     this.terminating = true;
   }
@@ -496,7 +528,7 @@ WorkerHandler.prototype.terminate = function (force, callback) {
 WorkerHandler.prototype.terminateAndNotify = function (force, timeout) {
   var resolver = Promise.defer();
   if (timeout) {
-    resolver.promise.timeout = timeout;
+    resolver.promise.timeout(timeout);
   }
   this.terminate(force, function (err, worker) {
     if (err) {