+++ src/types.js
@@ -4,13 +4,22 @@
  * @property {number} [maxWorkers]
  * @property {number} [maxQueueSize]
  * @property {'auto' | 'web' | 'process' | 'thread'} [workerType]
+ * @property {number} [workerTerminateTimeout]
  * @property {*} [forkArgs]
  * @property {*} [forkOpts]
  * @property {number} [debugPortStart]
  * @property {number} [maxJobsPerWorker]
+ * @property {Function} [onCreateWorker]
+ * @property {Function} [onTerminateWorker]
  */
 
 /**
  * @typedef {Object} ExecOptions
  * @property {(payload: any) => unknown} [on]
+ * @property {Object[]} [transfer]
+ */
+
+/**
+ * @typedef {Object} WorkerRegisterOptions
+ * @property {(code: number | undefined) => Promise | void} [onTerminate]
  */